Revert "[C-2772] Unit Tests For Core Methods (#54)"
This reverts commit 033432ab0e1a60118eb3398c5cc91ea7923d1f9d.
