+++ packages/react-inbox/src/components/Inbox/Bell.tsx
@@ -51,9 +51,7 @@ const Bell: React.ForwardRefExoticComponent<{
         onMouseEnter={onMouseEnter}
         ref={ref}
       >
-        {hasUnreadMessages && (
-          <div data-testid="unread-badge" className="unread-badge" />
-        )}
+        {hasUnreadMessages && <div className="unread-badge" />}
         <svg
           width="22"
           height="26"
