+++ packages/react-provider/babel.config.js
@@ -4,6 +4,12 @@ module.exports = {
     "@babel/transform-runtime",
     "babel-plugin-styled-components",
     "transform-class-properties",
+    [
+      "babel-plugin-inline-import",
+      {
+        extensions: [".css"],
+      },
+    ],
     ["babel-plugin-react-remove-properties", { properties: ["data-testid"] }],
     process.env.NODE_ENV !== "test" && [
       "babel-plugin-root-import",
